backend/requirements.txt
pandas
langchain-openai
openai
python-dotenv
chromadb
langchain-core
numpy
pandas
pdfplumber
pytest
sentence-transformers